requirements.txt
streamlit
langchain
langchain-groq
langchain-text-splitters
PyPDF2
python-dotenv